Almost done with JDBCs

+++ src/main/java/com/techelevator/model/Coffee.java
@@ -10,7 +10,7 @@ public class Coffee {
 	public long getCoffeeId() {
 		return this.coffeeId;
 	}
-	public void setCoffeeId(int coffeeId) {
+	public void setCoffeeId(long coffeeId) {
 		this.coffeeId = coffeeId;
 	}
 	public String getCoffeeName() {
